api/requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.0.0
tqdm>=4.66.0
google-generativeai>=0.3.0
Pillow>=10.0.0
chromadb==0.4.22
python-dotenv==1.0.0
langchain>=0.1.0
openai==1.12.0
tokenizers>=0.15.0
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.1
gunicorn>=21.2.0
pre-commit==3.5.0
tiktoken==0.5.2
boto3>=1.34.0
python-multipart>=0.0.6
google-cloud-storage>=2.0.0
grpcio>=1.59.0
protobuf>=4.23.0
ruff>=0.3.0
